Initial commit
Migrating from a private repo and squashing commits so I can
give Brian co-author credit. This version mostly works: it
generates code for types, events, and commands, and I have
smoke tested it by importing the generated package and also
testing with a trivial wrapper script that sends a command
to Chrome.

Co-authored-by: Brian Mackintosh <bcmackintosh@gmail.com>