Fix generator type
In d318f7ff I added type annotations for generators, but I got the
types wrong. The middle type is "send" type, which is "dict" for all
of our commands. The last type is "return" type, which is the one that
should contain a CDP type. This fixes a boatload of mypy errors.