Fix multi-return expressions
Some commands return multiple values. We have handled this so far by
returning a dictionary, but I noticed that the dictionary can contain
primitive types (e.g. str) and we were trying to parse a response like
`str.from_response(response['foo'])` which is not a valid method for
str. In this commit, primitive types are constructed directly, i.e.
`str(response['foo'])`.

This fixes a whole lot of mypy errors.