Properly serialize dataclasses
When a command function creates its JSON dictionary, it needs to call
asdict() on each dataclass object to turn it into a dictionary. But it
doesn't know which types are simple types and which types are
dataclasses, so to get the job done quickly I've added an is_subclass()
guard to each parameter.