Hack in support for optional response values.
Some commands can return multiple values, and some of those values are
optional. In lieu of full-blown support for optional values, I quickly
hacked in a dict.get() to replace the [] key access.

Note: this causes a bunch of mypy errors. I will reverse this change
when I add support for optional parameters, which should fix all of the
mypy errors at the same time.