Refactor build script
This is step 1 of a series of commits to refactor the build script. It
is currently very messy and unmaintainable, because the first time we
wrote we wanted to read the protocol spec and emit code in a single pass.
As we implemented all of the features, this approach became more and
more burdernsome, but we stuck with it in the interest of finishing a
completely working version as quickly as possible.

The new approach is going to be parsing the JSON up front, processing
it in memory, and then emitting Python code. This approach should be
much cleaner and will enable us to move from 4 files per CDP domain to
just 1 file per domain.

In this first commit, I've just added data classes to read in the CDP
spec. No changes to codegen, yet.