Flesh out the code for generating modules
Now that types are implemented, I added some code to the CdpDomain object
so that it can generate the code for a module file. (It outputs "TODO"
for commands and events, which are not implemented yet.)

I didn't write a test for this because it depends on functionality that
isn't finished, but also I think we'll notice quickly if the module
generation code is broken. (If I'm wrong about this, I'll add a test later
on.)