Rework dependencies
As we learned during our prototype, the dependencies that CDP declares in its
spec are insufficient for generating type-safe Python code. I thought maybe
this would have changed since all external types are quoted, but it doesn't. So
the new approach is to ignore the dependencies that CDP declares and compute
them ourselves. This is covered by an updated test.