Generate argument/return types from type hints #10
I added the sphinx-autodoc-typehints module, which uses typehint
information to label argument types and return types, plus hyperlinks
those types. The result is much more readable and usable documentation.
There is unfortunately a bug with Python 3.7.5 dataclasses and type
annotations that produces some warnings during builds, but this should
be resolved in 3.7.6 (as well as 3.8.1).